refactor(core-p2p): simplify handlers and remove dead code
