+++ __tests__/unit/core-p2p/socket-server/versions/internal/handlers/utils.test.ts
@@ -1,28 +1,10 @@
 import "../../../../mocks/core-container";
 import { eventEmitter } from "../../../../mocks/event-emitter";
 
-import {
-    emitEvent,
-    getUsernames,
-} from "../../../../../../../packages/core-p2p/src/socket-server/versions/internal/handlers/utils";
+import { emitEvent } from "../../../../../../../packages/core-p2p/src/socket-server/versions/internal";
 import { createPeerService } from "../../../../../../helpers/peers";
-import { delegates } from "../../../../../../utils/fixtures/testnet/delegates";
 
 describe("Internal handlers - utils", () => {
-    describe("getUsernames", () => {
-        it("should call blockchain forceWakeup", async () => {
-            const expectedData = {};
-            delegates.forEach(delegate => {
-                expectedData[delegate.publicKey] = "coolUsername";
-            });
-
-            const usernames = await getUsernames();
-            expect(usernames).toEqual({
-                data: expectedData,
-            });
-        });
-    });
-
     describe("emitEvent", () => {
         it("should emit the event provided", () => {
             const req = {
@@ -32,7 +14,7 @@ describe("Internal handlers - utils", () => {
                 },
             };
 
-            emitEvent(createPeerService().service, req);
+            emitEvent({ service: createPeerService().service, req });
 
             expect(eventEmitter.emit).toHaveBeenCalledWith(req.data.event, req.data.body);
         });
