+++ packages/core-p2p/src/socket-server/versions/index.ts
@@ -1,0 +1,5 @@
+import * as internal from "./internal";
+import * as peer from "./peer";
+import * as utils from "./utils";
+
+export { internal, peer, utils };
