--- packages/core-p2p/src/socket-server/versions/internal/handlers/network.ts
@@ -1,5 +1,0 @@
-import { P2P } from "@arkecosystem/core-interfaces";
-
-export const getNetworkState = async (service: P2P.IPeerService) => ({
-    data: await service.getMonitor().getNetworkState(),
-});
