--- packages/core-p2p/src/socket-server/versions/internal/index.ts
@@ -1,6 +1,0 @@
-export * from "./handlers/blockchain";
-export * from "./handlers/blocks";
-export * from "./handlers/network";
-export * from "./handlers/rounds";
-export * from "./handlers/transactions";
-export * from "./handlers/utils";
